+++ addons/hw_posbox_homepage/controllers/main.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import json
@@ -17,7 +16,7 @@ from odoo.modules.module import get_resource_path
 
 from odoo.addons.hw_drivers.main import iot_devices
 from odoo.addons.hw_drivers.tools import helpers
-from odoo.addons.web.controllers import main as web
+from odoo.addons.web.controllers.home import Home
 
 _logger = logging.getLogger(__name__)
 
@@ -46,7 +45,7 @@ six_payment_terminal_template = jinja_env.get_template('six_payment_terminal.htm
 list_credential_template = jinja_env.get_template('list_credential.html')
 upgrade_page_template = jinja_env.get_template('upgrade_page.html')
 
-class IoTboxHomepage(web.Home):
+class IoTboxHomepage(Home):
     def __init__(self):
         super(IoTboxHomepage,self).__init__()
         self.updating = threading.Lock()
