+++ addons/portal/controllers/web.py
@@ -2,25 +2,25 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import http
-from odoo.addons.web.controllers import main
+from odoo.addons.web.controllers.home import Home as WebHome
 from odoo.http import request
 
 
-class Home(main.Home):
+class Home(WebHome):
 
     @http.route()
     def index(self, *args, **kw):
         if request.session.uid and not request.env['res.users'].sudo().browse(request.session.uid).has_group('base.group_user'):
             return request.redirect_query('/my', query=request.params)
-        return super(Home, self).index(*args, **kw)
+        return super().index(*args, **kw)
 
     def _login_redirect(self, uid, redirect=None):
         if not redirect and not request.env['res.users'].sudo().browse(uid).has_group('base.group_user'):
             redirect = '/my'
-        return super(Home, self)._login_redirect(uid, redirect=redirect)
+        return super()._login_redirect(uid, redirect=redirect)
 
     @http.route()
     def web_client(self, s_action=None, **kw):
         if request.session.uid and not request.env['res.users'].sudo().browse(request.session.uid).has_group('base.group_user'):
             return request.redirect_query('/my', query=request.params)
-        return super(Home, self).web_client(s_action, **kw)
+        return super().web_client(s_action, **kw)
